Moved exevute function to a different file
